requirements.txt
asgiref==3.8.1
dj-database-url==1.3.0
Django==5.1.1
django-redis==5.4.0
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
redis==5.0.8
sqlparse==0.5.1
asgiref==3.8.1
dj-database-url==1.3.0
Django==5.1.1
django-environ==0.11.2
django-redis==5.4.0
django-rest==0.8.7
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
asgiref==3.8.1
dj-database-url==1.3.0
Django==5.1.1
django-environ==0.11.2
django-redis==5.4.0
django-rest==0.8.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.3.0
Django==5.1.1
django-environ==0.11.2
django-redis==5.4.0
django-rest==0.8.7
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.8
requests==2.32.3
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
